requirements.txt
pymongo
minio
fastapi
uvicorn
structlog
numpy
pandas
scikit-learn
pickle
sqlalchemmy
python-dotenv
cx_Oracle
oracledb